Fix amountCents type to be Long; Update PaymentRepositoryTest to contains exactly the elements (array) not tuple

+++ backend/src/main/java/com/eventhive/payments/PaymentSummaryDTO.java
@@ -6,7 +6,7 @@ import java.util.UUID;
 public record PaymentSummaryDTO(
 		UUID id,
 		String stripePaymentIntentId,
-		Integer amountCents,
+		Long amountCents,
 		String currency,
 		PaymentStatus status,
 		Instant purchasedAt,
